Move models file to user.py
