+++ app/api/services/users/create_user.py
@@ -2,7 +2,7 @@ from uuid import uuid4
 
 from api.db.models.user_models import User, UserRole
 from api.route.api_context import ApiContext
-from api.route.models.model import UserParams, UserResponse
+from api.route.models.user import UserParams, UserResponse
 
 
 # TODO: take in the UserParams... and return a User object...
